+++ audiocontext/include/interface.h
@@ -2,6 +2,6 @@
 
 void _audiocontext_init(audiocontext_config*);
 
-void _audiocontext_feed(float**, size_t, size_t, int);
+void _audiocontext_feed(float**, int, int, int);
 
 void _audiocontext_close();
